removed mac support
currently, the mac support for iOS apps that are not specifically mac catalyst enabled is limited to apple silicon, and i do not have an apple silicon device to test with. additionally, mac catalyst apps are based on UIKit, not SwiftUI, therefore making porting this app to mac very difficult, and it is not something that is a priority for me at the moment.